+++ shared/feature/clients/src/commonMain/kotlin/com/yellowtrack/platform/feature/clients/ClientsModule.kt
@@ -17,6 +17,6 @@ val clientsModule =
         viewModel { ClientsViewModel(get(), get(), get(), get(), get()) }
 
         viewModel { (clientId: ClientId) ->
-            ClientDetailsViewModel(clientId, get(), get(), get(), get())
+            ClientDetailsViewModel(clientId, get(), get(), get(), get(), get())
         }
     }